Replace "uncompress" with "decompress" in code and documentation
Updated method names and documentation to use "decompress" instead
of "uncompress" for greater clarity and correctness when referring
to the removal of file compression.

All incorrect uses of “uncompress” in the code and documentation
have been replaced. We use “decompress” to describe the action of
removing compression from a file. “uncompressed” should refer
only to the state of a file that is no longer compressed.

References: BAR-770

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>