Cleanup staging directory after its operation is done
Adds a cleanup step in MainRecoveryExecutor to the staging directory
after it was used after the next operation. So, if we have
operations = [A,B,C], after B we remove staging dir of A, after C we
remove dir of B.

- Adds a method to RecoveryOperation that handles the deletion of the
staging directory.
- Adds a property called `staging_path` to RecoveryOperation to build
the name of the staging_path per operation.
- Adds logic to MainRecoveryExecutor to call the method that cleans
up a staging directory of a previous operation, as explained above.

References: BAR-842

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>