Remove legacy recovery executor classes
Removed the legacy recovery executor classes `IncrementalRecoveryExecutor`
(and `SyntheticBackupInfo`, as it only existed to assist this class)
and `TarballRecoveryExecutor`, as their functionality is now implemented
by the new recovery operation classes `CombineOperation` and
`DecompressOperation` together with the new `MainRecoveryExecutor`.

As a consequence, the `recovery_executor_factory` function has been
updated to only return the `SnapshotRecoveryExecutor` or the
`MainRecoveryExecutor`, depending on the type of backup being restored,
as the legacy executors are no longer available.

References: BAR-844

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>