Remove danger notes about overwriting pgdata and tablespaces when restoring
It removes the danger notes about overwriting pgdata and tablespaces
when restoring a backup to a node where there is already a postgres
cluster with the same pgdata and tablespaces locations Or when
restoring a backup locally.

In BAR-787 #478, we have updated this behavior to raise an error and
quit if the recovery target is an existing, non-empty directory.

References: BAR-845
References: BAR-787

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>