Fix oversight of `_get_command_interface` for RsynCopyOperation
This fixes a bug found when restoring a backup remotely with
`staging_location=local`. The bug happens because the
`RsyncCopyOperation` is using the `UnixLocalCommand` interface
incorrectly to copy the backup when it should be using the
`UnixRemoteCommand` to copy the backup remotely.  For the
`RsynCopyOperation` what really matters when getting the correct
command interface (`UnixLocalCommand` or `UnixRemoteCommand`) is
the `remote_command` only, without considering the `staging_location`.

If `remote_command=None`, we know its a local restore, so
`staging_location` is not even set.
If remote_command="ssh pghost", we know its a remote restore and if
`staging_location=local`, the rsync command still needs to use the
remote interface to copy the backup into the remote node.

References: BAR-849
References: BAR-803

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>