Remove retry machanism for pg_combinebackup
The retry handler method passed to `PgCombineBackup` had the wrong signature,
which caused errors whenever called.

Instead of fixing the handler, after internal discussion, it was decided
to remove the retry mechanism for `pg_combinebackup` entirely, as:

1. There are not that many variables involved in combining backups, so
   if something failed it is likely because it was supposed to and a
   retry is not going to produce a different result.
2. Similar operations, e.g. decompression and decryption, which are also
   part of the recovery process do not have any retry mechanism at all.
4. The general rule throughout the code seems to be to only have retry
   mechanisms for things that are network bound, such as `rsync` and
   `pg_basebackup`.

Due to that, this commit removes the whole logic of retrying
`pg_combinebackup` from the code.

References: BAR-852

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>