Set default value of "local" for staging_location during recovery
This commit sets a default value of "local" for the `staging_location`
configuration option.

While testing local recovery of compressed backups, we identified an
issue: for such backups, `staging_location` is unnecessary (the
recovery is already local), and `staging_path` is also not needed
since decompression is performed directly into the final destination.

After reviewing the recovery logic, we determined that assigning a
default value of "local" to `staging_location` aligns better with user
expectations. It’s generally preferable for Barman to handle decompression,
decryption, and/or backup combination locally before transferring the
result to the remote node, rather than performing those operations on the
remote system.

We also set a default for `staging_path='/tmp'` at runtime so we do not
break compatibility with deprecated options.

References: BAR-850

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>