Fix restore failure on mixed backup chains by skipping unnecessary operations
When restoring an incremental compressed backup where its parent
was not compressed, the following error was thrown out:

Traceback (most recent call last):
  File "/home/vagrant/Repositories/edb/barman-internal/barman/recovery_executor.py", line 2592, in _decompress_backup
    compressor = compressors[compression](self.cmd)
KeyError: None

This commit addresses the bug by introducing an additional
`_should_execute` check within the `_execute_on_chain` method for
each backup in the chain. This ensures that the operation is
only performed on backups that actually require it.

The method signatures have been slightly changed because destination
is a common argument for all operation methods that inherit from the
parent class and it seems reasonable to add it in the front after
backup_info so we have this distinction of what is required and what
is optional. Also, as we need the destination in the
`_execute_on_chain`, this seems a good approach.

It also adds a logic to prepare directories, move/copy backups that
were not operated on and link tablespaces, so we can safely cleanup
the staging paths in between operations, instead of having to move
the cleanup of all to the end.

References: BAR-863

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>