Copy backup_manifest in DecompressOperation
Fixed an issue in the `DecompressOperation` where the backup manifest
was not being copied to the next staging directory. The manifest file is
required by subsequent a `CombineOperation` that might come later.
This is only required when `DecompressOperation` is not the last
operation in the recovery process.

The copy of the manifest file itself was implemented using a new `copy`
method added to the `fs` module, which uses the `cp` command. Using rsync
here was avoided because it doesn't support copying files between two
locations on the same remote server. E.g. there might be a case where
`DecompressOperation` is executed on a remote server and has to copy the
manifest file from `/remote/server/staging/barman-decryption/backup_manifest` to
`/remote/server/staging/barman-decompress/backup_manifest`. Doing that via
rsync results in errors like the one below:

EXCEPTION: {'ret': 1, 'out': '', 'err': 'The source and destination cannot both be remote.\nrsync error: syntax or usage error (code 1) at main.c(1420) [Receiver=3.2.3]\n'}

To avoid that, a simple `cp` command turned out to be easier and more reliable.

References: BAR-853

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>