Add explicit recommendations regarding usage of NFS with Barman
This commit makes more visible some recommendations regarding
the use of NFS as the storage for backups and WALs.

First we make sure the word NFS appears in the configuration
option for barman_lock_directory so it's easier to spot.

Second, we revive a section from the old docs that wasn't
migrated regarding recommendations on Backups. This shares
information on how the development team tests and certain
recommendations on how to use Barman together with an NFS
storage.

References: BAR-862

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>