Map gzip and bzip2 to pybzip and pybzip2 compression
Compression options such as `gzip`, `bzip2` and `pigz` uses subprocesses
to perform compression and decompression. This is not desirable in
Barman anymore due to the performance impact of forking new processes
and depending on external tools which might not be available at all times.

That said, this commit maps the `gzip` and `bzip2` compression options
to their equivalent `pygzip` and `pybzip2`, respectively, which are
Python-based implementations of the same compression algorithms.

This change is part of the effort to move away from subprocess-based
compression in Barman.

References: BAR-878

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>