Add '-p JOBS' to barman-wal-restore command when --get-wal is `True`
Adds '-p JOBS' to barman-wal-restore, JOBS being the number of jobs
chosen by the user via `-p` flag or `parallel_jobs` configuration
option.

Add two use cases to the one existing unit test of
`test_generate_recovery_conf_pre12`. There are two tests (one pre
postgres 12 and other post). The restore_command code block is not
affected by the version so we do not need to test in both places.

Slightly modifies the `parse_recovery_conf` test helper to skip
lines the do not have a '='. For the restore_command test cases,
a 'comment' is appended before the actualy command so we must skip
the parsing of this comment.

References: BAR-857

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>