Fix paralallel jobs of barman-wal-restore
There has been a regression when introducing the `--keep-compression`
flag in a previous release, which caused the parallel jobs to not
work as expected.

What happened is that the previous implementation of parallelism
relied on invoking multiple subprocesses (`get-wal` via SSH +
decompression, if needed) and piping the output of the first to the
second.

When introducing the `--keep-compression` flag, `barman-wal-restore`
started to handle decompression methods that were not done via
subprocess calls but rather through Python libraries. This
completely broke the parallelism, as the pipes could no longer
work asynchronously, as the decompression without subprocesses
is synchronous and requires the input content to be available
before it can start processing.

That said, to continue supporting parallelism in `barman-wal-restore`,
we had to completely change the way it is implemented.

This commit introduces a new approach for parallel jobs which uses
`multiprocessing` processes. A lot of refactoring was done to
accommodate this change. But the main idea is that now all the execution
of `RemoteGetWal` happens synchronously, and parallelism is achieved
by running multiple `RemoteGetWal` instances in separate processes.

Closes #1080

References: BAR-865

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>