Add validation to pg_combinebackup
Adds a validation between pg_combinebackup tool version and the
version of postgres for the backup being combined.

If they mismatch, we error out with a message. This is needed
because pg_combinebackup will only work if the postgres version of
the backup matches the version of the tool.

References: BAR-886

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>