Modify JSON output of barman diagnose command
This commit rearranges the JSON output of the barman diagnose command
by:

1. It replaces vars(self) by self.KEYS when fetching SERVER CONFIG
information so we do not include ServerConfig class attributes. This
class attributes are in fact GLOBAL configuration.
2. Those class attributes are not present in the server config json
block of the barman diagnose anymore. Instead, now it is included
in the global key of the JSON ouput since those values are options
that can be set by the user in the barman.conf file. They are added
in the global_config_to_json under Config class.
3. The name key was removed from the server config json block because
it is already present as the key of a server.
4. The msg_list attribute was moved to a higher level in the JSON tree
outside of the config block, but inside the server.

Models were not changed as it is already sibling of servers and global
config.

Finally, fix broken tests.

References: BAR-134

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>