Refactor `get_wal_conninfo` to ouput info message of applied overrides
This commit refactors the  `get_wal_conninfo` without changing the logic
or fallbacks of each configuration. The refactor was implemented so
we can better format the output info message and to accommodate
the new parameter `check_strategy` so we only display that message
in non check commands.

The `check_strategy` parameter was added because if we do not have
a way to omit the display of that message when doing a barman check,
the output will be odd and ugly with a loose message in between the
check rows. For example:

```
barman check streaming-backup-server
Server streaming-backup-server:
        PostgreSQL: OK
        superuser or standard user with backup privileges: OK
        PostgreSQL streaming: OK
        wal_level: OK
        PostgreSQL server is standby: OK
        Primary server is not a standby: OK
        Primary and standby have same system ID: OK
Configuration overrides applied: 'streaming_conninfo' overridden with 'host=postgres user=streaming_barman' and 'conninfo' overridden with 'host=postgres user=barman dbname=postgres'
        has monitoring privileges (WAL streaming): OK
        PostgreSQL streaming (WAL streaming): OK
```

References: BAR-442

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>