Add --combine-mode option to barman restore command
Added a `--combine-mode` option to the `barman restore` command, which can be used
to specify a copy mode for `pg_combinebackup` when combining incremental backups
during a restore. Available modes are `copy`, `link`, `clone` and `copy-file-range`
    which are equivalent to the `pg_combinebackup` options `--copy`, `--link`,
`--clone`, and `--copy-file-range`, respectively. An equivalent `combine_mode`
configuration option has also been added to specify it at the server level.

This feature allows for greater optimization of the restore process
based on the underlying filesystem capabilities.

References: BAR-870

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>