Refactor test_connectivity test for cloud_wal_archive
This commit refactors how we handle the test_connectivity check and
bucket creation.

1. Barman will not automatically create the target bucket when running
`barman-cloud-wal-archive`. The user must create the bucket before.
2. The `test_connectivity` check that tries to "access" the bucket
via `head_bucket` to inspect if the bucket exists will only be run
if the user provides the `-t/--test` flag when running
`barman-cloud-wal-archive`. This gives flexibility to the user by
keeping the option of testing connectivity while keeping the
`-t/--test` behavior.

References: BAR-895

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>