Add and refactor unit tests based on the changes of test connectivity
1. An interface method was created
   `verify_cloud_connectivity_and_bucket_existence` which is responsible
to call `test_connectivity` (this assigns `True` or `False` to the
`bucket_exists` attribute) and also checks whether this attribute was
assigned a value (bucket exists) or not (bucket does not exist). This
method will raise an exception depending on the outcome.
2. All existing unit tests were modified or removed because we already
test two possible outcome errors with this new method.

References: BAR-908

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>