Fix pg_combinebackup not found in PATH when staging location is remote
As reported in issue #1110, when recovering an incremental backup
with `staging_location` being `remote`, the restore was failing with
an error like:

```
    ERROR: /srv/pgsql-17.2/bin/pg_combinebackup not in PATH
```

This is because, besides the command being searched in the remote `PATH`,
it was also being searched in the local `PATH`, which is not necessary
when the staging location is remote. So the restore ends up failing
if `pg_combinebackup` is not installed in the local machine as well.

The issue happens because `PgCombineBackup` (any `Command` actually) checks
its availability in the local `PATH` when instantiated. And since
remote execution also requires instantiating the `PgCombineBackup`
the check was performed even when not necessary.

The fix is to skip the local `PATH` check via a new parameter `skip_path_check`
added to the `Command.__init__` method.

References: BAR-903
Closes: #1110

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>