Fix bug in get_path_device_number function
A bug was found in the `get_path_device_number` utils function that
was using the os module. The function was using the local paths
independently of the restore type (local or remote). This is wrong
because this function in only used for remote restore and it should
be checking the paths in the remote node.

This commit changes the get_path_device_number to be called as a
Unix subcommand instead of using the os module. By doing this we
make use of the remote unix subcommand to check for the filesystem.

References: BAR-920

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>