Fix delete_under_prefix bug for cloud backups and WAL deletion
A bug was identified in the `delete_under_prefix` method when
deleting backups and WALs in the cloud, related to the workaround
for the `MissingContentMD5` error delivered in BAR-909.

When running:

barman-cloud-backup-delete -v --profile minio \
--endpoint-url http://3.95.10.87:9000 s3://minio-barman-backup \
barman-test -b 20251001T184831

the following error occurred:

2025-10-01 18:52:05,397 [29824] ERROR: Deletion of objects under barman-test/wals/0000000100000000/ failed with error code: "204"
2025-10-01 18:52:05,397 [29824] ERROR: Barman cloud backup delete exception: Command exited with non-zero status 4

This error was also present in the previous implementation.

This commit reverts the changes made for the batch deletion so that
the old implementation is used, while retaining individual WAL
deletion when a `MissingContentMD5` error is encountered.

References: BAR-922

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>