Handle non-UTF-8 bytes in file subprocess output
When archiving WAL files, Barman executes the `file` command to check if
the file is encrypted. However, there was a report of the `file` command
returning a `UnicodeDecodeError` exception, likely because the output
contained bytes that could not be decoded using UTF-8.
To prevent this, we modify the subprocess call to handle such bytes
gracefully using the `surrogateescape` error handler.

Closes: #1117

References: BAR-926

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>