Catch more non-compliant S3 object stores for object deletion
In commit f5c9804040e9f1e2aa6d0109a04b5e69f0f398a5 we increased the
scope of failures to delete objects to catch not only when the
error code was "MissingContentMD5", but also on the message that
is returned.

We've found that different object stores return different error
codes and also different error messages, making some still fail
to delete the objects with "content-MD5" error messages.

This commit increases the scope of error messages we catch, which
we hope will end the problems we've been having with non-compliant
object stores.

References: BAR-974

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>