Support the S3NS variant of GCP
S3NS is a new deployment of GCP, with operations delegated by Google to
Thales: https://www.s3ns.io/en

S3NS is officially supported by Google, but is totally separate from the
regular GCP, with notably its own API domains.

We're porting our platform to S3NS, which includes CNPG configured with
the Barman Cloud plugin for backups, which doesn't work there
right now because we can't configure the Google *universe*.

To connect to S3NS, official clients need to be configured with
a universe value, that overrides the base API domain.
`GOOGLE_CLOUD_UNIVERSE_DOMAIN` is the official environment variable used
to target S3NS: https://documentation.s3ns.fr/docs/overview/tpc-key-differences#key_differences_for_developers

Unfortunately Barman uses the v1 client that doesn't read this variable,
and only accepts the universe as part of the `client_options` field
in the constructor, so we have to pass it in like this.